Refactor game engine class
