+++ pekascape/game.py
@@ -1,68 +1,52 @@
 import asyncio
 import random
+from typing import Self
 
 from pekascape.element import Food, Monster, Player, Weapon
 from pekascape.environment.environment import GridMap
-from user_control.user_control import UserControl
+from pekascape.user_interface.user_control import UserControl
 
 
-class CustomGame:
+class GameEngine:
     """
     This should represent a custom game, but currently also serves as input-action parser
     # TODO: refactor this class, split it into input parser and game class
     """
 
-    def __init__(self) -> None:
-
-        self.world_size = None
-        self.monsters = None
-        self.weapons = None
-        self.food = None
-
-        self.world = None
-
-        self.command_queue = asyncio.Queue()
-        self.command_semaphore = asyncio.Semaphore(1)
-        self.control = UserControl(self.command_queue, self.command_semaphore)
-
-        self._start_new()
-        self._get_config()
-        self._parse_config_values()
-        self._create_world()
-        self._populate_world()
-
-    def _start_new(self) -> None:
+    @classmethod
+    async def create_custom_game(cls) -> Self:
         print('hello there, you are starting new game of PeKaScape')
-        i = input("please type in your name:")
-        self.player_name = i
+        player_name = input("please type in your name:")
 
-    def _get_config(self) -> None:
         values = ["5*5", 50, 20, 20]
         for index, value in enumerate(['world size (in format of x*y)', 'monster count (int)', 'weapon count (int)', 'food count (int)']):
             values[index] = input(f"please, fill in the value for {value}:")
 
-        self.world_size, self.monsters, self.weapons, self.food = values
+        world_size, monsters, weapons, food = values
+        world_size = (int(x) for x in world_size.split('*'))
 
-    def _parse_config_values(self) -> None:
-        self.world_size = (int(x) for x in self.world_size.split('*'))
+        world_map = GridMap(*world_size)
 
-    def _create_world(self) -> None:
-        self.map = GridMap(*self.world_size)
+        # populate
+        player = Player(name=player_name, room=world_map.random_frame, health=500, attack=20, defence=20)
+        for _ in range(int(monsters)):
+            Monster(room=world_map.random_frame, attack=random.randint(10, 50))
 
-    def _populate_world(self) -> None:
-        print("Populating world with monsters, weapons and food")
+        for _ in range(int(weapons)):
+            Weapon.create_random(room=world_map.random_frame)
 
-        self.player = Player(name=self.player_name, room=self.map.random_frame, health=500, attack=20, defence=20)
-        for _ in range(int(self.monsters)):
-            Monster(room=self.map.random_frame, attack=random.randint(10, 50))
+        for _ in range(int(food)):
+            Food.create_random(room=world_map.random_frame)
 
-        for _ in range(int(self.weapons)):
-            Weapon.create_random(room=self.map.random_frame)
+        return cls(player)
 
-        for _ in range(int(self.food)):
-            Food.create_random(room=self.map.random_frame)
+    def __init__(self, player, world_map) -> None:
+        self.player = player
+        self.world_map = world_map
 
-        print("done")
+        self.command_queue = asyncio.Queue()
+        self.command_semaphore = asyncio.Semaphore(1)
+        self.control = UserControl(self.command_queue, self.command_semaphore)
 
     @staticmethod
     def _parse_action(action: str) -> tuple[str, str]:
